standard-api-response

+++ server/controllers/book.js
@@ -1,6 +1,7 @@
 import Book from "../models/Book.js";
 import Retailer from "../models/Retailer.js";
-import User from "../models/User.js";
+import { createSuccess } from "../utils/success.js";
+import { createError } from "../utils/error.js";
 
 export const createBook = async (req, res, next) => {
 
@@ -11,7 +12,7 @@ export const createBook = async (req, res, next) => {
         // Find the retailer by userId
         const retailer = await Retailer.findOne({ userId });
         if (!retailer) {
-            return res.status(404).json({ message: "Retailer not found" });
+            return res.status(404).json(createError(404, "Retailer not found."));
         }
 
         const newBook = new Book({
@@ -25,7 +26,7 @@ export const createBook = async (req, res, next) => {
         retailer.listedBooks.push(savedBook._id);
         await retailer.save();
 
-        res.status(201).json("Book has been created.");
+        res.status(201).json(createSuccess("Book has been created.", savedBook));
     } catch (err) {
         next(err);
     }
@@ -34,18 +35,28 @@ export const createBook = async (req, res, next) => {
 
 export const updateBook = async (req, res, next) => {
     try {
-        const updatedBook = await Book.findById(req.params.bookId);
+        const { id: userId, bookId } = req.params;
+        const updatedBook = await Book.findById(bookId);
 
         // Check if the book exists
         if (!updatedBook) {
-            return res.status(404).json({ message: "Book not found" });
+            return res.status(404).json(createError(404, "Book not found."));
+        }
+        const retailer = await Retailer.findOne({ userId });
+
+        if (!retailer) {
+            return res.status(404).json(createError(404, "Retailer not found."));
+        }
+
+        if (!retailer._id.equals(updatedBook.retailerId)) {
+            return res.status(401).json(createError(401, "Unauthorized."));
         }
 
         // Update the book
         updatedBook.set(req.body);
         const savedBook = await updatedBook.save();
 
-        res.status(200).json(savedBook);
+        res.status(200).json(createSuccess("Book has been updated.", savedBook));
     } catch (err) {
         next(err);
     }
@@ -53,21 +64,29 @@ export const updateBook = async (req, res, next) => {
 
 export const deleteBook = async (req, res, next) => {
     try {
-        const book = await Book.findById(req.params.bookId);
+        const { id: userId, bookId } = req.params;
+        const book = await Book.findById(bookId);
 
         // Check if the book exists
         if (!book) {
-            return res.status(404).json({ message: "Book not found" });
+            return res.status(404).json(createError(404, "Book not found."));
         }
 
+        const retailer = await Retailer.findOne({ userId });
+        if (!retailer) {
+            return res.status(404).json(createError(404, "Retailer not found."));
+        }
+        if (!retailer._id.equals(book.retailerId)) {
+            console.log(retailer._id, book.retailerId, retailer._id !== book.retailerId);
+            return res.status(401).json(createError(401, "Unauthorized."));
+        }
 
-        const retailer = await Retailer.findOne({ _id: book.retailerId });
         retailer.listedBooks.pull(req.params.bookId);
         await retailer.save();
 
         await Book.deleteOne({ _id: req.params.bookId });
 
-        res.status(204).json({ message: "Book has been deleted." });
+        res.status(204).json(createSuccess("Book has been deleted."));
 
 
 
@@ -79,7 +98,7 @@ export const deleteBook = async (req, res, next) => {
 export const getBook = async (req, res, next) => {
     try {
         const book = await Book.findById(req.params.id);
-        res.status(200).json(book);
+        res.status(200).json(createSuccess("Book found.", book));
     } catch (err) {
         next(err);
     }
@@ -88,7 +107,7 @@ export const getBook = async (req, res, next) => {
 export const getBooks = async (req, res, next) => {
     try {
         const books = await Book.find(req.query);
-        res.status(200).json(books);
+        res.status(200).json(createSuccess("Books found.", books));
     } catch (err) {
         next(err);
     }
@@ -97,9 +116,10 @@ export const getBooks = async (req, res, next) => {
 // Get all books listed by a retailer
 export const getRetailerBook = async (req, res, next) => {
     try {
+
         const retailer = await Retailer.findOne({ userId: req.params.id });
         const books = await Book.find({ retailerId: retailer._id });
-        res.status(200).json(books);
+        res.status(200).json(createSuccess("Retailer's books found.", books));
     } catch (err) {
         next(err);
     }
